Started examples.

+++ server.js
@@ -15,39 +15,59 @@ function renderCSS(callback){
 }
 
 /* Renders the JS */
-function renderElm(callback) {
-  var cmd = 'elm-make source/kitchensink.elm --output test.js';
+function renderElm(str) {
+  return function(callback) {
+    var cmd = `elm-make '${str}' --output test.js`;
+    exec(cmd, function(error, stdout, stderr) {
+      if (stderr) {
+        err = stderr.replace(/\n/g,"\\n")
+        callback(null, `console.error("${err}")`);
+      } else {
+        callback(null, fs.readFileSync('test.js', 'utf-8'))
+        fs.unlink('test.js')
+      }
+    });
+  }
+}
 
-  exec(cmd, function(error, stdout, stderr) {
-    callback(null, fs.readFileSync('test.js', 'utf-8'))
-    fs.unlink('test.js')
-  });
+function renderHtml(str) {
+  return `<html>
+      <head>
+        <title>Elm-UI Kitchesink</title>
+      </head>
+      <body>
+        <script src='${str}' type='application/javascript'>
+        </script>
+        <script>Elm.fullscreen(Elm.Main);</script>
+      </body>
+    </html>`
 }
 
 /* Serve HTML */
 router.get('/', function *(next) {
-  this.body = `<html>
-                <head>
-                  <title>Elm-UI Kitchesink</title>
-                </head>
-                <body>
-                  <script src='index.js' type='application/javascript'>
-                  </script>
-                  <script>Elm.fullscreen(Elm.Main);</script>
-                </body>
-              </html>`
+  this.body = renderHtml('index.js')
 })
 
 /* Serve JS */
 router.get('/index.js', function *(next) {
   this.type = 'text/javascript';
-  this.body = yield renderElm
+  this.body = yield renderElm('source/kitchensink.elm')
 })
 
 /* Serve CSS */
 router.get('/index.css', function *(next) {
   this.type = 'text/css';
   this.body = yield renderCSS
+})
+
+ router.get('/_examples/:id', function *(next){
+  this.type = 'text/javascript';
+  this.body = yield renderElm(`examples/${this.params.id}.elm`)
+})
+
+router.get('/examples/:id', function *(next){
+  this.type = 'text/html'
+  this.body = renderHtml(`/_examples/${this.params.id}`)
 })
 
 app
